Refactor statement handlers into separate modules
